+++ src/main/resources/lib/cache/find-paths-to-invalidate.ts
